refactor(anyharness): split session domain responsibilities
