+++ anyharness/crates/anyharness-lib/src/domains/reviews/runtime/events.rs
@@ -5,7 +5,7 @@ use super::super::model::{ReviewFeedbackJobRecord, ReviewRunStatus};
 use super::super::service::ReviewError;
 use super::ReviewRuntime;
 use crate::sessions::extensions::SessionTurnOutcome;
-use crate::sessions::prompt::PromptProvenance;
+use crate::sessions::prompt::provenance::PromptProvenance;
 use crate::sessions::runtime_event::RuntimeInjectedSessionEvent;
 
 impl ReviewRuntime {
