+++ anyharness/crates/anyharness-lib/src/live/sessions/actor/config/selection.rs
@@ -2,8 +2,9 @@ use agent_client_protocol as acp;
 
 use crate::live::sessions::actor::config::types::ConfigPurpose;
 use crate::live::sessions::actor::state::SessionStartupState;
+use crate::sessions::live_config::controls::option_matches_key;
 use crate::sessions::live_config::{
-    normalized_key_rank, option_matches_key, NormalizedControlKind, LEGACY_MODE_COMPAT_CONFIG_ID,
+    normalized_key_rank, NormalizedControlKind, LEGACY_MODE_COMPAT_CONFIG_ID,
 };
 pub(in crate::live::sessions::actor) fn pending_config_rank(
     startup_state: &SessionStartupState,
