+++ anyharness/crates/anyharness-lib/src/sessions/service/title.rs
@@ -1,0 +1,34 @@
+use super::{SessionService, UpdateSessionTitleError};
+use crate::sessions::model::SessionRecord;
+
+impl SessionService {
+    pub fn update_session_title(
+        &self,
+        session_id: &str,
+        title: &str,
+    ) -> Result<SessionRecord, UpdateSessionTitleError> {
+        let trimmed = title.trim();
+        if trimmed.is_empty() {
+            return Err(UpdateSessionTitleError::EmptyTitle);
+        }
+        if trimmed.chars().count() > 160 {
+            return Err(UpdateSessionTitleError::TitleTooLong(160));
+        }
+
+        let existing = self
+            .session_store
+            .find_by_id(session_id)
+            .map_err(UpdateSessionTitleError::Internal)?
+            .ok_or_else(|| UpdateSessionTitleError::SessionNotFound(session_id.to_string()))?;
+
+        let now = chrono::Utc::now().to_rfc3339();
+        self.session_store
+            .update_title(session_id, trimmed, &now)
+            .map_err(UpdateSessionTitleError::Internal)?;
+
+        let mut updated = existing;
+        updated.title = Some(trimmed.to_string());
+        updated.updated_at = now;
+        Ok(updated)
+    }
+}
