+++ scripts/max_lines_allowlist.txt
@@ -20,9 +20,6 @@ anyharness/crates/anyharness-lib/src/domains/mobility/service.rs 1534 existing A
 anyharness/crates/anyharness-lib/src/domains/plans/service.rs 653 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/domains/reviews/service/service_tests.rs 1519 existing AnyHarness oversized test file
 anyharness/crates/anyharness-lib/src/domains/runtime_config/service.rs 1057 existing AnyHarness oversized file
-anyharness/crates/anyharness-lib/src/sessions/live_config.rs 675 existing AnyHarness oversized file
-anyharness/crates/anyharness-lib/src/sessions/prompt.rs 1423 existing AnyHarness oversized file
-anyharness/crates/anyharness-lib/src/sessions/service.rs 645 existing AnyHarness oversized file
 anyharness/crates/proliferate-worker/src/commands/dispatcher.rs 1718 existing Proliferate worker oversized file
 anyharness/crates/proliferate-worker/src/commands/mapping.rs 954 existing Proliferate worker oversized file
 anyharness/sdk-react/src/hooks/sessions.ts 653 existing SDK React oversized file
